# [go/mysql] use sync.Pool for reusing bufio readers and writers