Readers recycle objects on Read if there are no buffered bytes. Writers recycle objects on Flush. Both recycle on Reset.
Maybe we should also recycle writers on Write in similar to Read manner.
/cc @danieltahara @sougou 
benchmark results:
```
benchmark                            old ns/op      new ns/op      delta
BenchmarkParallelShortQueries-8      3443           3439           -0.12%
BenchmarkParallelMediumQueries-8     19907          23057          +15.82%
BenchmarkParallelRandomQueries-8     1098953101     1077136194     -1.99%

benchmark                            old MB/s     new MB/s     speedup
BenchmarkParallelShortQueries-8      6.10         6.10         1.00x
BenchmarkParallelMediumQueries-8     823.82       711.25       0.86x

benchmark                            old allocs     new allocs     delta
BenchmarkParallelShortQueries-8      24             24             +0.00%
BenchmarkParallelMediumQueries-8     24             25             +4.17%
BenchmarkParallelRandomQueries-8     32292          31851          -1.37%

benchmark                            old bytes     new bytes     delta
BenchmarkParallelShortQueries-8      867           873           +0.69%
BenchmarkParallelMediumQueries-8     63241         69361         +9.68%
BenchmarkParallelRandomQueries-8     147128744     142483592     -3.16%

```