@danieltahara 2018-09-07
As a related question -- why don't we flush after sending all rows in a result set?

https://github.com/vitessio/vitess/blob/master/go/mysql/query.go#L590
https://github.com/vitessio/vitess/blob/master/go/mysql/server.go#L371