@LK4D4 2018-09-07
So, I really forgot to recycle on Read.
But with recycle on Read benchmarks looks like:
```
benchmark                            old ns/op     new ns/op     delta
BenchmarkParallelShortQueries-8      3407          3566          +4.67%
BenchmarkParallelMediumQueries-8     18171         28565         +57.20%
BenchmarkParallelRandomQueries-8     11452408      10863779      -5.14%

benchmark                            old MB/s     new MB/s     speedup
BenchmarkParallelShortQueries-8      6.16         5.89         0.96x
BenchmarkParallelMediumQueries-8     902.49       574.12       0.64x

benchmark                            old allocs     new allocs     delta
BenchmarkParallelShortQueries-8      24             24             +0.00%
BenchmarkParallelMediumQueries-8     25             28             +12.00%
BenchmarkParallelRandomQueries-8     83             58             -30.12%

benchmark                            old bytes     new bytes     delta
BenchmarkParallelShortQueries-8      979           1212          +23.80%
BenchmarkParallelMediumQueries-8     69074         89985         +30.27%
BenchmarkParallelRandomQueries-8     66810632      62952562      -5.77%
```
It's encouraging that random queries are so much better, but other benchmarks are weird.