@derekperkins 2018-09-07
Is it possible that garbage collection is causing the weird benchmarks?
https://blog.cloudflare.com/go-dont-collect-my-garbage/