@danieltahara 2018-09-07
@derekperkins we had that happen in the previous diff, and it's obviously even worse here (i.e. heap gets even smaller since we're pooling everything) https://github.com/golang/go/issues/27545