@danieltahara 2018-09-08
So one note of clarification here is that from our perspective this isn't a
perf/latency play -- it's a GC improvement and will help tail latencies.
The benchmarks are just to show no or acceptable levels of throughput
regression.

This is why I bring up the assymmetry between a client and a server. A
client is really only hurt by anything were doing here -- pooling is
overkill and introducing a second syscall reduces throughput. Most client
applications will have no more than a few hundred connections.

In contrast, a server has to pay the buffer cost for every inbound
connection, and this can number in the hundreds of thousands. If you've
been following along with the golang GC issue, this results in multi
hundred second stop the world sweeps (and high concurrent mark). Lowering
the buffer to 4kb will help a tad but doesn't avoid the fundamental problem
of effectively immortal allocations clogging up concurrent mark and
concurrent sweep. This is what pooling and removing buffers aims to solve.

Hence my propose solution is to have a forked read path for the server
where we just dont use bufio (your suggestion) at all and have two syscall,
leaving the buffering for client.

On Sat, Sep 8, 2018 at 2:08 PM Sugu Sougoumarane <notifications@github.com>
wrote:

> *@sougou* commented on this pull request.
> ------------------------------
>
> In go/mysql/conn.go
> <https://github.com/vitessio/vitess/pull/4186#discussion_r216140843>:
>
> > @@ -136,8 +135,8 @@ type Conn struct {
>  	ClientData interface{}
>
>  	// Packet encoding variables.
> -	reader   *bufio.Reader
> -	writer   *bufio.Writer
> +	reader   bufioReader
> +	writer   bufioWriter
>  	sequence uint8
>
>
> In spite of the benchmarks, I don't think the pooling will give us
> anything in real life. This is because most connections will generally be
> idle, and during that time, they will be possessing a buffer waiting on a
> read.
>
> If we went with the 4-byte approach, we'll need to perform some real-life
> benchmarks to evaluate the impact of the extra syscall. If we find that
> it's worth it, then we should not use bufio at all:
> All waiting reads can fetch into a 4-byte buffer, and then we can directly
> read the rest of the data into the target slice. There is no benefit to
> using bufio in this case.
>
> Another option is to reduce the size of bufio to 4K, which is what go
> recommends. With the removal of the other two buffers, the overhead of a
> connection would drop to 4K, which can be more easily accommodated.
>
> We should just look at dropping connBuffSize and use go's default anyways.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/vitessio/vitess/pull/4186#discussion_r216140843>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/ABQKuTmV5tfi3zIb76ICaFu7g-2aVh4Zks5uZDHmgaJpZM4Wd4Mr>
> .
>
-- 
Daniel

dktahara@gmail.com
(646) 397-6379