@danieltahara 2018-09-09
Looping in some offline conversation --

@sougou agrees in the assessment that the additional syscall per mysql packet could pose problems for latency sensitive clients, in particular folks at Slack. It would be great if we could have @demmer or @rafael (or whoever is appropriate) benchmark whatever client-side solution we come up with.

In light of the need to minimize impact on those using the connection as a mysql client (vs as a server), I'd propose forking the code paths to have a buffered path and an unbuffered path. A naive implementation of this would invite a lot of complexity -- casing out the header read or `ReadQueryResult`, but I believe there's a simpler way to achieve the same end: offer an additional constructor for the connection object `newConnServer`, which doesn't instantiate a buffered reader and uses the connection directly (I don't even think we need to bother with pooling the bufio readers, although we could do it if we want). This removes the buffer problem for a mysql listener/server while preserving existing behavior for clients.